add docblocks to ShopStaff class
